Move music2.models to music2.common.models
Also forgot
* put the classes from the link module into the __init__.py file
* create special handler for list[str]
* fix SubmissionHistory names in SeedHistory class in seed submodule
