+++ music2/common/models/genre.py
@@ -11,7 +11,7 @@ genre_relation table.
 import sqlalchemy
 from sqlalchemy import orm
 
-from music2.models import base
+from music2.common.models import base
 
 __all__ = ("Genre", "GenreRelation")
