+++ music2/common/models/link.py
@@ -11,7 +11,7 @@ song, or bracket entry.
 import sqlalchemy
 from sqlalchemy import orm
 
-from music2.models import base
+from music2.common.models import base
 
 __all__ = ("Link", "LinkRelation")
