+++ music2/common/models/match.py
@@ -18,7 +18,7 @@ import datetime
 import sqlalchemy
 from sqlalchemy import orm
 
-from music2.models import base
+from music2.common.models import base
 
 __all__ = ("Match", "MatchSeed")
