+++ music2/common/models/notification.py
@@ -10,7 +10,7 @@ import datetime
 import sqlalchemy
 from sqlalchemy import orm
 
-from music2.models import base
+from music2.common.models import base
 
 __all__ = ("Notification",)
