+++ music2/common/models/participant.py
@@ -14,7 +14,7 @@ import datetime
 import sqlalchemy
 from sqlalchemy import orm
 
-from music2.models import base
+from music2.common.models import base
 
 __all__ = ("Participant", "ParticipantNickname")
