+++ music2/common/models/project.py
@@ -19,7 +19,7 @@ import uuid
 import sqlalchemy
 from sqlalchemy import orm
 
-from music2.models import base
+from music2.common.models import base
 
 __all__ = ("BracketSelection", "Config", "Project")
