+++ music2/common/models/round.py
@@ -13,7 +13,7 @@ import datetime
 import sqlalchemy
 from sqlalchemy import orm
 
-from music2.models import base
+from music2.common.models import base
 
 __all__ = ("Round",)
