+++ music2/common/models/submission.py
@@ -11,7 +11,7 @@ either by the user via a command or by an admin.
 import sqlalchemy
 from sqlalchemy import orm
 
-from music2.models import base
+from music2.common.models import base
 
 __all__ = ("Submission", "SubmissionHistory")
